Move HandlesPageFilesystemTest to unit test namespace
